% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance.indices.R
\name{balance.indices}
\alias{balance.indices}
\title{Computes Colles-like, Sackin and cophenetic balance indices of a phylogenetic tree.}
\usage{
balance.indices(tree, norm = FALSE)
}
\arguments{
\item{tree}{a single phylogenetic tree. It can be introduced as a string in the Newick format, as a 'phylo' object (\code{ape} package) or as an 'igraph' object (\code{igraph} package).}

\item{norm}{a logical object indicating if the indices should been normalized or not.}
}
\value{
A numeric vector with the three computed indices of the tree: 
\code{Colless-like}, \code{Sackin} and \code{Cophenetic} values.
}
\description{
Computes Colles-like, Sackin and cophenetic balance indices of a phylogenetic tree.
}
\details{
The Colless-like index is the generalization of the Colless' index for non-binary trees.

The Sackin's index is computed as the sum of the number of ancestors for each leave of the tree.

Cophenetic index is computed as the sum of the depths of the least common ancestor (LCA) of every pair of leaves.
}
\examples{
#From a newick string
balance.indices("(1,2,3,4,5);")
balance.indices("(1,(2,(3,(4,5))));")

#From a phylo object
require(ape)
random.tree = rtree(5,rooted=TRUE)
balance.indices(random.tree)

#An example of a tree generated by the alpha-gamma model (igraph object)
a.g.tree = a.g.model(5,0.5,0.3)
balance.indices(a.g.tree)

#All of them can be normalized (between 0 and 1)
balance.indices("(1,2,3,4,5);",norm=TRUE)
balance.indices("(1,(2,(3,(4,5))));",norm=TRUE)
balance.indices(random.tree,norm=TRUE)
balance.indices(a.g.tree,norm=TRUE)

}
\references{
A. Mir, F. Rossello, L.Rotger, A Colless-like balance index for multifurcating phylogenetic trees.

A. Mir, F. Rossello, L.Rotger, A new balance index for phylogenetic trees. Math. Biosc. 241 (2013).

M. J. Sackin, "Good" and  "bad" phenograms. Sys. Zool, 21 (1972), 225-226.
}
\author{
Lucia Rotger
}
