% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colles.like.index.R
\name{colless.like.index}
\alias{colless.like.index}
\title{Computes the Colless-like index of a tree}
\usage{
colless.like.index(tree, f.size = "ln", diss = "MDM", norm = FALSE)
}
\arguments{
\item{tree}{a single phylogenetic tree. It can be introduced as a string in the Newick format, as a "phylo" object (\code{ape} package) or as an "igraph" object (\code{igraph} package).}

\item{f.size}{by default is \code{f(n)=ln(n+e)}. It can be set as "ln" (default) or "exp" (\code{f(n)=exp(n)}). It aso can be a function defined by user, in this case the index can not be normalized.}

\item{diss}{by default the dissimilarity is \code{MDM}. It can also be set as "var" (the sample variance) or "sd" (the sample standard deviation). It aso can be a function defined by user, in this case the index can not be normalized.}

\item{norm}{a logical object indicating if the indices should been normalized or not.}
}
\value{
A numeric value.
}
\description{
Computes the Colless-like balance index of a phylogenetic tree.
}
\details{
The Colless-like index is the generalization of the Colless' index for non-binary trees. 

By default, the f-size function is \code{f(n)=exp(n)} and the dissimilarity is the mean deviation from the median (MDM). 
It is possible to change them by specifying it with the parameters \code{f.size} and \code{diss}, with "exp" the f-size would be \code{f(n)=exp(n)}, and with "var" (or "sd") the dissimilarity would be the sample variance (or the sample standard deviation).
It is also possible to set a new function for both parameters, see "References".
}
\examples{
#' #From a newick string
colless.like.index("(1,2,3,4,5);")
colless.like.index("(1,(2,(3,(4,5))));")

#From a phylo object
require(ape)
random.tree = rtree(5,rooted=TRUE)
colless.like.index(random.tree)

#An example of a tree generated by the alpha-gamma model (igraph object)
a.g.tree = a.g.model(5,0.5,0.3)
colless.like.index(a.g.tree)

#All of them can be normalized (between 0 and 1)
colless.like.index("(1,2,3,4,5);",norm=TRUE)
colless.like.index("(1,(2,(3,(4,5))));",norm=TRUE)
colless.like.index(random.tree,norm=TRUE)
colless.like.index(a.g.tree,norm=TRUE)

#Changing f-size 
colless.like.index(a.g.tree,f.size="exp")

#Changing dissimilarity
colless.like.index(a.g.tree,diss="var")
colless.like.index(a.g.tree,diss="sd")

#Changing both
colless.like.index(a.g.tree,f.size="exp",diss="var")
colless.like.index(a.g.tree,f.size="exp",diss="sd")

}
\references{
A. Mir, F. Rossello, L.Rotger, A Colless-like balance index for multifurcating phylogenetic trees.
}
\author{
Lucia Rotger
}
