\name{CollessLike-package}
\alias{CollessLike-package}
\alias{CollessLike}
\docType{package}
\title{
\packageTitle{CollessLike}
}
\description{
\packageDescription{CollessLike}
}
\details{

%The DESCRIPTION file:
\packageDESCRIPTION{CollessLike}
\packageIndices{CollessLike}
%~~ An overview of how to use the package, including the most ~~
%~~ important functions ~~
}
\author{
\packageAuthor{CollessLike}

Maintainer: \packageMaintainer{CollessLike}
}
\references{ Chen, B., Ford, D., Winkel, M., A new family of Markov branching trees: the alpha-gamma model. \emph{Electr. J. Probab}. \bold{14} (2009), 400-430. MR2480547

A. Mir, F. Rossello, L.Rotger, A Colless-like balance index for multifurcating phylogenetic trees.
 
A. Mir, F. Rossello, L.Rotger, A new balance index for phylogenetic trees. Math. Biosc. 241 (2013).
 
M. J. Sackin, "Good" and  "bad" phenograms. Sys. Zool, 21 (1972), 225-226.
%~~ Literature or other references for background information ~~
}
%~~ Optionally other standard keywords, one per line, from file ~~
%~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
\examples{
# An example of a tree generated by the alpha-gamma model (igraph object)
 a.g.tree = a.g.model(5,0.5,0.3)
# To compute its percentile in the distribution
# distribution(a.g.tree,0.5,0.3,db.path=getwd())
 
# It is possible to do a new simulation of indices 
# ('Repetitions' set as 100 for a fast example)
  indices.data = indices.simulation(5,0.5,0.3,100)
  distribution(a.g.tree,0.5,0.3,set.indices=indices.data)
}
