% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{distribution}
\alias{distribution}
\title{Computes the percentile in the alpha-gamma distribution of the indices of the given tree}
\usage{
distribution(tree, alpha = NA, gamma = NA, set.indices = NULL,
  new.simulation = FALSE, repetitions = 1000,
  legend.location = "topright", cex = 0.75, percentile.plot = FALSE,
  db.path = getwd())
}
\arguments{
\item{tree}{a single phylogenetic tree. It can be introduced as a string in the Newick format, as a "phylo" object (\code{ape} package) or as an "igraph" object (\code{igraph} package).}

\item{alpha}{parametrer of the alpha-gamma model, between 0 and 1.}

\item{gamma}{parametrer of the alpha-gamma model, between 0 and alpha.}

\item{set.indices}{if \code{NULL}(default) the indices data is taken from stored data or from a new simulated data (See "Details"). If not, it must be a 3-column data.frame with the three balance indices (Colles-like, Sackin, Cophenetic). See \code{\link{indices.simulation}}.}

\item{new.simulation}{if \code{FALSE}(default) the indices data it could be from a data.frame introduced by the user or a data.frame from our database. If it is \code{TRUE}, a new indices data set is computed. See \code{\link{indices.simulation}}.}

\item{repetitions}{the number of trees to generate in case a new simulation is done.}

\item{legend.location}{location where the legend is going to be placed. See "Details".}

\item{cex}{expansion factor of the legend. See "Details".}

\item{percentile.plot}{if \code{TRUE} plots the percentile plot of the indices. If it is \code{FALSE}(default) , then a distribution plot is represented.}

\item{db.path}{by default is the actual working directory. It should be changed if the data base is going to be used and it is located in a different directory.}
}
\value{
A numeric vector with the three percentiles.
}
\description{
Computes the percentile in the alpha-gamma distribution of the indices (Colless-like, Sackin and Cophenetic) of the given tree. It also plots all the three distributions highlighting where are the indices of the given tree or it plots a percentile plot with the percentiles of the tree.
}
\details{
Two plots are available: one with the acumulated percentiles of the indices (\code{percentile.plot=FALSE}), and the other with the distribution (\code{percentile.plot=TRUE}).

The stored data available has been calculated for a number of leaves between 3 and 50. For each of them, the parameters are set as \code{alpha} in \{0,0.1,0.2,...,1\} and \code{gamma} in \{0,0.1,0.2,...,\code{alpha}\}. If the introduced parameters are not in the list, a new computation is done with them and a new dataset of trees is generated, and computed its indices. The number of trees generated can be modified by the parameter \code{repetitions} (see \code{\link{indices.simulation}} for more information). This computation may take some time, therefore you can computate it separately with \code{\link{indices.simulation}}, save its value and then call this function  by setting it as the parameter \code{set.indices}.

The legend is placed with the \code{graphics} function \code{legend()}, so its location can be specified by setting \code{legend.position} to a single keyword from the list \code{"bottomright"}, \code{"bottom"}, \code{"bottomleft"}, \code{"left"}, \code{"topleft"}, \code{"top"}, \code{"topright"}, \code{"right"} and \code{"center"}. 
The expansion factor for the legend is controlled by the parameter \code{cex}, by default \code{cex=1}. See \code{\link[graphics]{legend}}.
}
\examples{
#If it is need, to specify the location of the database
#folder=".../CollesLikeDataBase/" 
##If not,
folder=getwd()
## Different ways to introduce the tree
#From a newick string
distribution("(1,2,3,4,5);",0.5,0.3,db.path=folder)
distribution("(1,(2,(3,(4,5))));",0.5,0.3,db.path=folder)

#From a phylo object
require(ape)
random.tree = rtree(5,rooted=TRUE)
distribution(random.tree,0.5,0.3,db.path=folder)

#An example of a tree generated by the alpha-gamma model (igraph object)
a.g.tree = a.g.model(5,0.5,0.3)
distribution(a.g.tree,0.5,0.3,db.path=folder)

## Different indices data
# From our data base
distribution(a.g.tree,0.5,0.3,db.path=folder)

# From a data.frame generated by 'indices.simulation'
# ('Repetitions' set as 10 for a fast example)
indices.data = indices.simulation(5,0.5,0.3,10)
distribution(a.g.tree,0.5,0.3,set.indices=indices.data)

# Allow the function to do a new generation of data and compute their indices
distribution(a.g.tree,0.5,0.3,new.simulation=TRUE,repetitions=10)
# WARNING! it might take a long time, it depends on the parameters 
# 'n' (number of leaves) and 'repetition' (number of repetitions)

}
\references{
Chen, B., Ford, D., Winkel, M., A new family of Markov branching trees: the alpha-gamma model. \emph{Electr. J. Probab}. \bold{14} (2009), 400-430. MR2480547

A. Mir, F. Rossello, L.Rotger, A Colless-like balance index for multifurcating phylogenetic trees.

A. Mir, F. Rossello, L.Rotger, A new balance index for phylogenetic trees. Math. Biosc. 241 (2013).

M. J. Sackin, "Good" and  "bad" phenograms. Sys. Zool, 21 (1972), 225-226.
}
\seealso{
\code{\link[graphics]{legend}}, \code{\link{indices.simulation}}, \code{\link{balance.indices}}
}
\author{
Lucia Rotger
}
