% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sackin.index.R
\name{sackin.index}
\alias{sackin.index}
\title{Computes the Sackin index of a tree}
\usage{
sackin.index(tree, norm = FALSE)
}
\arguments{
\item{tree}{a single phylogenetic tree. It can be introduced as a string in the Newick format, as a "phylo" object (\code{ape} package) or as an "igraph" object (\code{igraph} package).}

\item{norm}{a logical object indicating if the indices should been normalized or not.}
}
\value{
numeric value.
}
\description{
Computes the Sackin index of balance of a phylogenetic tree.
}
\details{
The Sackin's index is computed as the sum of the number of ancestors for each leave of the tree.
}
\examples{
#From a newick string
sackin.index("(1,2,3,4,5);")
sackin.index("(1,(2,(3,(4,5))));")

#From a phylo object
require(ape)
random.tree = rtree(5,rooted=TRUE)
sackin.index(random.tree)

#An example of a tree generated by the alpha-gamma model (igraph object)
a.g.tree = a.g.model(5,0.5,0.3)
sackin.index(a.g.tree)

#All of them can be normalized (between 0 and 1)
sackin.index("(1,2,3,4,5);",norm=TRUE)
sackin.index("(1,(2,(3,(4,5))));",norm=TRUE)
sackin.index(random.tree,norm=TRUE)
sackin.index(a.g.tree,norm=TRUE)

}
\references{
M. J. Sackin, "Good" and  "bad" phenograms. Sys. Zool, 21 (1972), 225-226.
}
\author{
Lucia Rotger
}
