% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a.g.model.R
\name{a.g.model}
\alias{a.g.model}
\title{Generates a random tree}
\usage{
a.g.model(n, alpha, gamma)
}
\arguments{
\item{n}{the number of leaves in the tree.}

\item{alpha}{parametrer of the alpha-gamma model, between 0 and 1.}

\item{gamma}{parametrer of the alpha-gamma model, between 0 and alpha.}
}
\value{
An igraph object that is the ford's tree created.
}
\description{
Generates a random tree according to the alpha-gamma model
}
\examples{
 
tree = a.g.model(10,0.8,0.1) 
# plot(tree,layout=layout.reingold.tilford(tree,root=which(degree(tree,mode="in")==0)))

tree = a.g.model(5,0.5,0.3)
# plot(tree,layout=layout.reingold.tilford(tree,root=which(degree(tree,mode="in")==0)))

}
\references{
\insertRef{alphagamma}{CollessLike}
}
\author{
Lucia Rotger
}
