% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cophen.index.R
\name{cophen.index}
\alias{cophen.index}
\title{Computes the cophenetic index of a phylogenetic tree}
\usage{
cophen.index(tree, norm = FALSE)
}
\arguments{
\item{tree}{a single phylogenetic tree. It can be introduced as a string in the Newick format, as a 'phylo' object (\code{ape} package) or as an 'igraph' object (\code{igraph} package).}

\item{norm}{a logical object indicating if the indices should been normalized or not.}
}
\value{
A numeric value.
}
\description{
Computes the cophenetic index of balance of a phylogenetic tree.
}
\details{
The cophenetic index is computed as the sum of the depths of the least common ancestor (LCA) of every pair of leaves.
}
\examples{
#' #From a newick string
cophen.index("(1,2,3,4,5);")
cophen.index("(1,(2,(3,(4,5))));")

#From a phylo object
require(ape)
random.tree = rtree(5,rooted=TRUE)
cophen.index(random.tree)

#An example of a tree generated by the alpha-gamma model (igraph object)
a.g.tree = a.g.model(5,0.5,0.3)
cophen.index(a.g.tree)

#All of them can be normalized (between 0 and 1)
cophen.index("(1,2,3,4,5);",norm=TRUE)
cophen.index("(1,(2,(3,(4,5))));",norm=TRUE)
cophen.index(random.tree,norm=TRUE)
cophen.index(a.g.tree,norm=TRUE)

}
\references{
A. Mir, F. Rossello, L.Rotger, A new balance index for phylogenetic trees. Math. Biosc. 241 (2013).
}
\author{
Lucia Rotger
}
